Moved the test case to where it should be, based on the CONTRIBUTING documentation.
